requirements.txt
# -------------------------
# Core utilities / config
# -------------------------
python-dotenv>=1.0.1,<2
PyYAML>=6.0.1,<7
pydantic>=2.7,<3
orjson>=3.10,<4

# -------------------------
# RSS + web ingest / parsing
# -------------------------
feedparser>=6.0.11,<7
httpx>=0.27,<1
beautifulsoup4>=4.12,<5
lxml>=5.2,<6
dateparser>=1.2,<2
python-dateutil>=2.9,<3

# -------------------------
# Data / ML basics
# -------------------------
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
scipy>=1.12,<2

# -------------------------
# Deep learning (TextCNN + DRN)
# -------------------------
torch>=2.3,<3
torchvision>=0.18,<1

# -------------------------
# NLP helpers
# -------------------------
regex>=2024.5.15
spacy>=3.7,<4
sentence-transformers>=3.0,<4

# -------------------------
# Graph + clustering (incident merging)
# -------------------------
networkx>=3.2,<4
hdbscan>=0.8.39
rapidfuzz>=3.9,<4

# -------------------------
# Memory / retrieval (NTM-lite)
# -------------------------
faiss-cpu>=1.8.0,<2

# -------------------------
# API
# -------------------------
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1

# -------------------------
# Testing / dev
# -------------------------
pytest>=8.2,<9
pytest-cov>=5.0,<6
ruff>=0.5,<1
black>=24.4,<25
